Added methods to views and urls
